Reflow tune selection to be more pythonic
Variables in Python are not unreachable when they drop out of scope,
which means that variables do not need to be defined before a try/except
block, but can be done so inside it. Further, tighten the exception to
be IndexError rather than the generic Exception, and correct a typo in
the comment above the block.