Move song definitions into a YAML file
In the interests of starting to separate code and data, shift the
information about the tunes into a YAML file, which is then parsed into
a list. This also slightly changes the keys for parseability by humans.

Furthermore, un-urlencode the filenames by hand, and have Python do it
for us by using the quote() function.